+++ src/main/java/com/krazytop/leagueoflegends/nomenclature/ItemNomenclature.java
@@ -20,7 +20,7 @@ public class ItemNomenclature extends Nomenclature {
     private int baseGold;
     private int totalGold;
     private List<String> tags;
-    private Map<String, Double> stats;
+    private Map<String, Float> stats;
     @JsonAlias("into")
     private List<String> toItems;
     @JsonAlias("from")
